支持修改密码+展示喜欢图片+修改BUG+美化展示

+++ graduation project/app/deletePic.php
@@ -2,11 +2,15 @@
 	include("config.php");
 	
 	$pic_id = $_POST['pic_id'];
-
-	$sql = "delete from gp_pic where id='".$pic_id."'";
-	$result = mysqli_query($link, $sql);
+	$pic_url = $_POST['pic_url'];
 	
-	echo $pic_id;
-	mysqli_close($link);
+	$upload_path = "../".$pic_url;
+		
+	if(unlink($upload_path)){
+		$sql = "delete from gp_pic where id='".$pic_id."'";
+		$result = mysqli_query($link, $sql);echo $pic_id;
+		mysqli_close($link);
+	}
+	
 
 ?>
